+++ trunk/プログラム/Game/Object/GameScene/PlayerTime.cpp
@@ -96,8 +96,8 @@ void PlayerTime::RenderObject()
 	gd1.position = Vector2(m_x, m_y);
 	gd1.textureRec2 = Rect(0,0,128,32);
 	gd1.textureRec1 = Rect(0,32,128,64);
-	gd1.max = m_player.GetPlayerParameter().GetMaxPlayerTime();
-	gd1.point = m_player.GetPlayerParameter().GetPlayerTime();
+	gd1.max = (float)m_player.GetPlayerParameter().GetMaxPlayerTime();
+	gd1.point = (float)m_player.GetPlayerParameter().GetPlayerTime();
 	gd1.texture1 = TEXTUREID_HP;
 	gd1.texture2 = TEXTUREID_HP;
 	m_device.GetGraphicDevice().Render( gd1 );
